Fix for file names with space in them
Use the %q syntax for [fmt][1] in order to create a quoted and escaped
string. This way we can both copy and write files which have a space in
them.

Disclosure: Technically it is not entirely correct to use %q here
since Go's quoting rules aren't identical to those of a shell, but
it's good enough.

[1]: https://golang.org/pkg/fmt/