Update README.md to use CopyFromFile instead of CopyFile
CopyFile copies the entire file into memory which is a huge footgun for a lot of applications. It caused my test machine to OOM several times before I figured it out. Might be good to modify CopyFile to automaically check if it's a file to avoid this issue all together