Fixes issue #79: do not close user SSH connections
The issue was fixed in the way proposed in #79, an additional field
called `CloseHandler` was added to the `Client` struct. This handler can
either be `EmptyHandler` which is equivalent to a no-op, or a
`CloseSSHClient` which closes the passed SSH client when executed.

The `EmptyHandler` is used by default unless `Client.Connect()` is
called to establish the SSH connection. The reasoning is that whenever
someone passes their own `ssh.Client` to the library, `Connect()` will
not be called, thus signaling that we should not manage the
`ssh.Client`.

To ensure the correctness of this fix, two additional test cases were added
(1) `TestUserSuppliedSSHClientDoesNotClose` which creates an `scp`
client using a existing `ssh.Client` and ensures that the client is not
closed by attempting to create a new session from it.
(2) `TestSSHClientNoLeak` which uses `Connect()` to establish the SSH
connection and ensures that its underlying `ssh.Client` is no longer
functioning using the same mechanism as (1).